refactor(theme): rename barrel inspector_theme.dart → theme.dart
Align the barrel's filename with its directory (theme/theme.dart). Pure
rename plus import-path updates across the 12 referencing files; the
barrel content and all token classes are unchanged.
